+++ Sunrise.Shared/Extensions/Scores/ScoreExtensions.cs
@@ -246,7 +246,7 @@ public static class ScoreExtensions
 
         if ((int)score.GameMode != beatmap.ModeInt || (int)score.Mods > 0)
         {
-            var recalculateBeatmapResult = await calculatorService.CalculateBeatmapPerformance(session, score.BeatmapId, (int)score.GameMode, score.Mods);
+            var recalculateBeatmapResult = await calculatorService.CalculateBeatmapPerformance(session, score.BeatmapId, score.GameMode, score.Mods);
 
             if (recalculateBeatmapResult.IsFailure)
             {
